Change default number handling to float64
Decimals are cursed and usually not needed. We do lossy conversions by default in other places too (interval)

+++ cli/src/postgres_cloner.rs
@@ -99,7 +99,7 @@ pub fn default_settings() -> SchemaSettings {
 		json_handling: SchemaSettingsJsonHandling::Text, // DuckDB doesn't load JSON converted type, so better to use string I guess
 		enum_handling: SchemaSettingsEnumHandling::Text,
 		interval_handling: SchemaSettingsIntervalHandling::Interval,
-		numeric_handling: SchemaSettingsNumericHandling::Decimal,
+		numeric_handling: SchemaSettingsNumericHandling::Double,
 		decimal_scale: 18,
 		decimal_precision: 38,
 	}
